stm32_i2c: correct standard mode min bus frequency
Prior this change a typo set the standard mode min bus frequency
to 8kHz whereas it should be 80kH (80% of 100kHz).

This change explicitly sets the min/max bus frequency values
as 80% and 120% of the nominal frequency to overcome such issues.

This change also adds an explicit reference URL to the I2C
specification referred from an inline comment.

Change-Id: Iec4bee4fe77dcda6f411948750986c85ede7e1c1
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Reviewed-on: https://gerrit.st.com/126293
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>
Tested-by: Lionel DEBIEVE <lionel.debieve@st.com>
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>