stm32mp1: deactivate IO compensation in case of STOP mode
Disable IO compensation on A7 side in case of STOP mode.
SYSCFG clock has to be managed at this level, to ensure
access to the SoC interface registers when entering/exiting STOP mode.
This is done in the corresponding low_power service.

This change maps SoC SYSCFG interface registers as non-secure and
defines SYSCFG clock in the clock driver. Without futher change,
shared resource defines the clock as non-secure.

Change-Id: Ie3b075b3bfad45a0f1b489f369bbea4be8766df9
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/150129
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@st.com>