stm32mp1: pm: fix mpu/mcu/pll3/pll4 resume
When OP-TEE expects to enter CSTOP and CSTANDBY state, it is possible
that SoC only reaches STOP state since MCU subsystem prevented to reach
a deeper state. In that case, MPU/MCU/PLL3/PLL4 clocks context may be lost
and are not restored since system was not woken through TF-A that is
expected to restore all PLLs.

This change ensures PLL3 and PLL4 are restored in such case. A change
in pll_start() also ensures that already running PLLs are not
restarted. Prior this change was an issue since we disable PLL output
clocks P/Q/R when enabling PLL control bit.
Note that stm32mp1_set_clkdiv() local function as been added.

Change-Id: I27eb96c4b755cb338320bfb37a60ba0f0761dc95
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/160831
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>