plat-stm32mp1: optimize resident memory size
Decrease number of 4kB translation by 1 as 4 tables is sufficient
to map Core. Note that this assumes client will not request
mapping of more than 2MByte at a time.

Decrease Core heap size from 64kByte to 48kByte in order to save
memory for the pager. This configuration is not mandated and
platform configuration can override this value.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>