core: stm32mp1: rpoc pta: Add initialize service state
Add service init to initialize the hardware and context.

The PTA session can be open and closed several times.
In this case forcing the MCU RAM to secure context usage on
the session open is not valid. We also need to ensure that
the hardware is in expected state.

Default disable CONFIG_RPROC_PTA in platform configuration to prevent
MCU RAMs be secure when OP-TEE service are not used.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@st.com>
Change-Id: I28f9e61ca864b5047d7349eb2ee33b609fecc841
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/178976
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@st.com>
Tested-by: Etienne CARRIERE <etienne.carriere@st.com>