ta: remoteproc: keep TA alive
Add TA_FLAG_INSTANCE_KEEP_ALIVE flag to keep the TA alive.
This is mandatory to not loose the context when a first
session is used by U-boot to boot the remote proc firmware
and a second one is open on Linux kernel boot to attach to the
remote firmware.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@st.com>
Change-Id: I985a6dcf2492afd11f6939e284259fd1225eba44
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/178977
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@st.com>
Tested-by: Etienne CARRIERE <etienne.carriere@st.com>