plat-stm32mp1: correct IO compensation disabling
In stm32mp1_syscfg_disable_io_compensation(), to disable the IO
compensation cell, we have to set the corresponding bit in
SYSCFG_CMPENCLRR register, instead of clearing the bit in SETR register.
Force the clock enable during the disable_io as the clock is not secured
and may have been disabled.

Signed-off-by: Yann Gautier <yann.gautier@st.com>
Signed-off-by: Lionel Debieve <lionel.debieve@st.com>
Change-Id: I634102a8a2aa8094c82f43593d327a7a93478975
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/188746
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@st.com>