core: add stm32_tamp driver
The tamp driver manages the anti-tamper detection circuit.
Sensitive data in the device are protected from external attack by the
anti-tamper detection circuit. There are internal tampers (link with
internal state) and external tamper (link with one or two external pins)
that may raise an tamper event if some state are detected.
External tamper can be configured for edge detection, or level detection
with or without filtering, or active tamper which increases the security
level by auto checking that the tamper pins are not externally opened or
shorted.

This driver provides also an API to configure backup registers permission
access and lock (and make usable) the SAES BHK key.

Change-Id: I7ca4b8a1db7bdd77b4b51094e2f6e7b9db31b488
Signed-off-by: Nicolas Toromanoff <nicolas.toromanoff@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/203890
Tested-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>
Reviewed-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/209900
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>