core: verify size of allocated shared memory
Makes sure that normal world cannot change the size of allocated shared
memory, resulting in a smaller buffer being allocated.

This reintroduces cc6bc5f94210 ("core: verify size of allocated shared
memory") which was lost with Fixes: 2786f1438fc8 ("core: thread:
separate old SMC interface handling").

Since then we have a separate set of functions to deal with RPC when
communicating with FF-A. A corresponding size check is added for the
FF-A version of thread_rpc_alloc().

The change is based on Jens Wiklander patches [1] and [2],
backported to OP-TEE 3.9.0 with minor adaptation based on
prior changes.

Link: [1] 4ed45027f507 ("core: verify size of allocated shared memory")
Link: [2] ae2a9cfce320 ("core: fix shared memory buffer rpc allocation")

Change-Id: I729bbfd3d7d3fb5b6b20ee7b98c5c485867e2a00
Co-developped-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/225184
Tested-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>