drivers: stm32_iwdg: implementation of independent watchdog
Add of an independent watchdog, which offers a high safety level, thanks
to its capability to detect malfunctions due to software or hardware
failures. It is clocked by an independent clock and stays active if the
main clock fails.

For the list of features, refer to the reference manuals at:
https://wiki.st.com/stm32mpu/wiki/STM32MP15_resources

Depending on the mode (hardware or software), it can be started
at power-on/reset or by writing a value in the IWDG_KEY register
(see reference manual).

Config should be specified in the device-tree so stm32_iwdg_setup
function can set the watchdog properly.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
Change-Id: I562c997efc0581add1147c2608ba478a990ccad1