drivers: stm32_rng: force RNG clock under 750KHz
Implement an algorithm that gets the rate of the RNG clock and
apply the correct value in CLKDIV field in RNG_CR register to
force the RNG clock rate to be under 750KHz.

Remove cr as it is useless to read and keep the value of the register
before configuring RNG IP, whatever the platform.

Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>
Change-Id: I98a95731415c9136c48269b0276f39230cc8eb60