clk: stm32mp1_clk: skip always-on clocks in stm32mp1_clock_resume()
Gateable clocks, listed in stm32mp1_clk_gate[] table, get resumed in
stm32mp1_clock_resume() function, when returning from Standby state.
If a clock is always-on, like RTC, it doesn't get a refcount. So the
resume function tries to disable it. Such clocks should then be skipped
in the gate resume loop.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I3694ffbb36665592ff8d236bb123bd8735ab8e13
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/230810
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>