stm32_rng: prevent masking interrupt when HW RNG is busy
Change API functions reading RNG random byte to not mask interrupts
while pooling or waiting for some hardware state. This change modifies
the RNG error concealment sequence and the loops reading random bytes.

With the new implementation, data are read by burst of max size the
RNG FIFO size and interrupt are unmasked when waiting for the
FIFO the be filled. This is implemented by the helper function
get_rng_bytes_relaxed().

Change-Id: Ie8485757c72676a7d623360863c95e0ead7aaf4d
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/218060
Tested-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>