drivers: stm32_gpio: add secure configuration for gpios
This change adds security support for GPIOS. A bank of GPIO now has a
secure support and configuration.

Secure support is defined in the device tree. If a GPIO bank is defined
as secure, the secure configuration is read through st,protreg device tree
property and is applied during probe.

Secure configuration is kept during power management sequences.

Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>
Change-Id: I2b34100916fbe8be4b68f3cbc1f3e6211a47e52d